Size.exe
Added size.exe
Its a quick program that I made that asks for the dimensions of a
rectangle and it draws it. Helpful for figuring out how big something
will be like pictures or margins. To open it, go to the folder where you
saved this github project and it will be there.